+++ doc.go
@@ -36,7 +36,7 @@ Source code and other details for the project are available at GitHub:
 
 Current Version
 
-0.0.16
+0.0.17
 
 Installation
 
@@ -390,4 +390,4 @@ Examples:
 package golog // import "github.com/kataras/golog"
 
 // Version is the version string representation of the "golog" package.
-const Version = "0.0.16"
+const Version = "0.0.17"